Keep a self-play round's planes in half precision
The planes are the round: 370,000 decisions of 97 by 34 is 4.9 GB in
float32, and gathering them allocates the whole array while the blocks
are still alive, so the peak is twice that. Every number in an
observation is a count in unary, a fraction of a discard span, or a
score over fifty thousand; none needs more than three digits, and the
network reads them in bfloat16 anyway. Half precision halves the round
and halves the gather's peak with it, which is what was left after the
loop stopped carrying two rounds: eighteen gigabytes of private memory,
eight of it resident, with the machine paging and a generation taking
six minutes instead of two and a half. Each minibatch becomes float32
as it crosses to the card.

+++ neural/train.py
@@ -189,7 +189,10 @@ def main() -> None:
                 if picks.numel() < 2:
                     continue
                 drawn = picks.cpu()
-                planes = observations[drawn].to(device)
+                # Half precision on the host, float32 on the card: the
+                # tower's weights are float32, and autocast takes it from
+                # there.
+                planes = observations[drawn].to(device).float()
                 seen = oracle[drawn].to(device).float()
                 with torch.autocast("cuda", dtype=torch.bfloat16, enabled=args.amp):
                     logits, value, guessed, oracle_value = learn(planes, legal[picks], seen)
